Add explicit casts to fix warnings.

+++ ios/AirImagePicker/AirImagePicker.m
@@ -640,7 +640,7 @@ void AirImagePickerContextInitializer(void* extData, const uint8_t* ctxType, FRE
 {
     // Register the links btwn AS3 and ObjC. (dont forget to modify the nbFuntionsToLink integer if you are adding/removing functions)
     NSInteger nbFuntionsToLink = 6;
-    *numFunctionsToTest = nbFuntionsToLink;
+    *numFunctionsToTest = (uint32_t)nbFuntionsToLink;
     
     FRENamedFunction* func = (FRENamedFunction*) malloc(sizeof(FRENamedFunction) * nbFuntionsToLink);
